requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
chainmap==1.0.3
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
combomethod==1.0.12
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
nulltype==2.3.1
options==1.4.10
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
pygame==2.3.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==7.3.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
